refactor: guestbook 패키지 분리
