+++ backend/forgather/src/main/java/com/forgather/domain/guestbook/repository/jpa/GuestJpaRepository.java
@@ -1,0 +1,9 @@
+package com.forgather.domain.guestbook.repository.jpa;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.forgather.domain.guestbook.model.Guest;
+import com.forgather.domain.guestbook.repository.GuestRepository;
+
+public interface GuestJpaRepository extends JpaRepository<Guest, Long>, GuestRepository {
+}
